[Consensus] consider tx processed either via consensus or checkpoint when submit (#18896)
## Description 

This PR introduces logic to consider a transaction as processed when one
of the following conditions is met:

1. It has been sequenced through consensus.
2. It has been executed via a checkpoint (state sync).

Under typical circumstances, transactions are expected to be marked as
processed primarily through the consensus path. However, the checkpoint
execution path proves beneficial when a node faces difficulties syncing
with the consensus DAG but can still follow the state sync. This
approach prevents transactions from remaining as inflight and, more
importantly, avoids redundant submission through the consensus path, as
these transactions are likely to be obsolete.

## Test plan 

CI/Private Testnet

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: