[execution] Update how verifier constants for signing are set (#19094)
## Description 

Since verifier constants for limits and metering for signing do not need
to be protocol versioned, a previous PR moved some of these constants
out of the protocol config. This PR takes that one step further by
allowing it to be set by the node config instead of being hardcoded into
the binary. The default is that these should remain unset in the node
config, but this gives the ability to easily change them later on.

## Test plan 

Make sure existing tests pass.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: