[Move] Update tests to use unique addresses for different modules in tests (#19262)
## Description 

Updates tests so that unique addresses are used on modules that are not
part of the same package (needed as part of the larger package loader
rewrite). Note that the `0x1`, `0x2` addresses are used by the stdlib
(etc.) so I bumped all addresses to start at `0x6`.

Tests should largely be unchanged. The one exception is some rewriting
in the `move-vm-integration-tests` as there were issues running tests,
and the fact the `CompiledModule::serialize` is test-only.

Once this lands, I'll rebase and land into the vm_2024 branch.

## Test plan 

Update existing tests, verify their outputs change as expected.