chore(graphql): Align all e2e tests on protocol 51 (#19402)
## Description

Some of these tests are on older versions because they landed around the
time of the last bump. Bringing everything in line with protocol 51.

## Test plan

```
$ cargo nextest run -p sui-graphql-e2e-tests
```

## Stack

- #19371 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: