[cli] Switch to `WaitForEffectsCert` for CLI (#19375)
## Description 

Added a new function `execute_tx` that is used only in the CLI that uses
`WaitForEffectsCert` instead of `WaitForLocalExecution` to align with
the deprecation of waiting for local execution.
Also added the polling for transaction like in
https://github.com/MystenLabs/sui/pull/18996/files

## Test plan 

Existing tests.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: The CLI switched to using `WaitForEffectsCert` in the quorum
driver for transaction execution. Due to this, it simulates
`WaitForLocalExecution` via polling as JSON RPC ignores `WaitForLocalExecution` requests now.
- [ ] Rust SDK:
- [ ] REST API: