[sui-proxy/timeouts+errors] (#19454)
## Description
add an on_error to the trace layer
add a timeout to all routes, configurable via env vars or using the
default.

## Test Plan
locally

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ x ] Nodes (Validators and Full nodes): nodes sending metrics will
have an enforced timeout of 20 seconds. If metrics cannot be sent in
that time, the sui-proxy metrics server will disconnect the client. This
will not affect the blockchain or any related activity. It only applies
to metrics transmission.
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: