subscriptions: reinstate EventFilter::Any (#19807)
## Description

Temporarily re-enable `EventFilter::Any` as it is still in use in some
places. This is a partial revert of #19617.

## Test plan

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [x] Nodes (Validators and Full nodes): Temporarily re-enable
`EventFilter::Any` as a kind of event subscription filter. Note that
subscriptions are deprecated. This means they are not officially
supported, nor actively maintained.
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: