[GraphQL] Serialize bcs bytes of `TransactionData` instead of `SenderSignedData` (#19768)
## Description 

This PR modifies which data is serialized to bcs for `TransactionBlock`
type. Instead of serializing to bcs `SenderSignedData`, which includes
signatures and intent message, it only serializes the transaction data.

## Test plan 

Existing tests

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: The `bcs` field of `TransactionBlock` has new data.
Instead of serializing to `bcs` from `SenderSignedData`, which includes
signatures and intent message, it only serializes the `TransactionData`
object. Note that this breaks the semantics compared to previous JSON RPC which serialized the intent and signatures. `TransactionData` only includes tx data, but no signatures or intent.
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: