[Consensus] fix amnesia recovery boot run (#19774)
## Description 

Currently if someone recovers their node within epoch `R` with a
snapshot that's from epoch `< R-1` , consensus will not start in amnesia
recovery mode in epoch R as the boot counter will have already been
incremented as node is trying to catch up from earlier epochs. This is
problematic as it defies the whole point of the automatic amnesia
recovery.

Instead on this PR we track consensus participation activity from
earlier epochs/run and only then we increment the boot counter.
Otherwise we keep the boot counter to `0` so we effectively enforce
amnesia recovery until we get to an epoch where the node is able to
participate. In this case participate means "able to have performed at
least one commit".

## Test plan 

CI/PT

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: