refactor(indexer-alt): pull out `pipeline` (#19933)
## Descrption

Separate the `pipeline` abstraction from the `handler` abstraction, and
put it into its own module. This is in preparation for:

- Splitting up the batching and committing processes. This will make it
possible to test committing multiple chunks in parallel, and it may be
possible to re-use some of this logic in...
- an `in_order::pipeline` to use for pipelines that need to write out
data strictly in checkpoint order and along checkpoint boundaries.

## Test plan

Existing tests and CI.

## Stack

- #19926 
- #19932 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: