rest: introduce protobuf as an acceptable content type
Given the BCS format is inherently un-evolvable (unless you utilize
enums) it makes a poor wire format as any changes made lead to
incompatibilities with existing clients.

Today the REST api uses JSON as the default, and human-readable, format
and BCS for the binary one. This PR introduces Protobuf as another
possible format type and converts the `list_checkpoints` api to support
requesting a page of checkpoints in a protobuf format. The plan is to
introduce support for protobuf to other apis and to phase out bcs in
most places.