indexer-alt: sum_obj_types pipeline (#20054)
## Description

The `sum_obj_types` sequential pipeline is the `sui-indexer-alt`
equivalent of `objects` and `objects_snapshot`. It is a `sum`-mary of
the object owner and type information as of a particular checkpoint (its
watermark).

It differs from `objects` and `objects_snapshot` in the following ways:

- It only contains owner and type information (no coin balance, content,
digest, or dynamic field information), but has a more "complete" set of
indices for this data.
- The type information has been standardised into the same style as used
by event indices.
- It uses the painter's algorithm to avoid processing redundant object
updates (process transactions in a checkpoint in reverse order).
- Updates to the underlying table are atomic (all writes are sent out
within a single transaction, along with a watermark update).
- It avoids redundant writes across checkpoints as well (e.g. if an
object is modified multiple times, only the last modification will be
turned into a DB update).

This PR does not include the logic to delay updates (necessary to
replicate the behaviour of `objects_snapshot`). This will be added at a
later date.

## Test plan

Ran the indexer on the first 1M checkpoints in mainnet (produced 2481
live objects by the end):

```
sui$ cargo run -p sui-indexer-alt --release --                                   \
  --database-url "postgres://postgres:postgrespw@localhost:5432/sui_indexer_alt" \
  --remote-store-url https://checkpoints.mainnet.sui.io                          \
  --last-checkpoint 1000000
```

## Stack

- #20050 
- #20051 
- #20052 
- #20053 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: