fix(graphql): loading Owner.dynamicObjectField.value (#20346)
## Description

Today, when we fetch child objects, we can supply a root version, which
is the version of the object in storage that eventually owns this
object. We can't use the object's direct parent version because if we
modify a child but not its parent in a transaction, then the parent's
version will not be updated if that parent is itself a child object.

When we know that we are fetching a dynamic field off an object, O, if O
was initialised with a root version, that is inherited by the dynamic
field query, but if one was not supplied, then we use the object's
version as the root version.

In the case of dynamic object fields, the dynamic field value is itself
a child object of the `Field` that contains its name (which is itself a
child object of some other object).

In the case of reading a dynamic object field off an owner address,
these details all come together to form a bug: The owner does not supply
a root version, so we use the version from the `Field`, the `Field` has
not been modified since it was first assigned, so that means we're using
the initial version of the field, and we end up fetching the initial
version of the value as well.

The fix is to have a special carve out for dynamic fields' root
versions: The dynamic field needs to remember the root version it was
queried with and propagate that when querying its value (in the case of
a dynamic object field).

## Test plan

Fixed E2E tests:

```
sui$ cargo nextest run -p sui-graphql-e2e-tests -- owner/root_version
```

And test the query that instigated this. The following query on testnet
previously returned the first version of the dynamic field value, and
now correctly returns the latest version:

```
{
  owner(
    address: "0x37c0e4d7b36a2f64d51bba262a1791f844cfd88f31379f1b7c04244061d43914"
  ) {
    dynamicObjectField(name: {type: "u64", bcs: "AAAAAAAAAAA="}) {
      value {
        ... on MoveObject {
          contents {
            json
          }
        }
      }
    }
  }
}
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: Fix a bug when reading a dynamic object field value from
an `Owner`, where the value returned would not be the latest version
when the `owner` was queried without a root/parent version supplied.
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: